+++ lib/language_pack/helpers/bundler_wrapper.rb
@@ -36,7 +36,7 @@ class LanguagePack::Helpers::BundlerWrapper
   include LanguagePack::ShellHelpers
 
   BLESSED_BUNDLER_VERSIONS = {}
-  # Heroku-22's oldest Ruby version is 3.1
+  # Heroku-22's oldest Ruby version is 3.1, but Ruby 2.6.6 is supported via heroku-20 fallback
   BLESSED_BUNDLER_VERSIONS["2.3"] = "2.3.25"
   BLESSED_BUNDLER_VERSIONS["2.4"] = "2.4.22"
   BLESSED_BUNDLER_VERSIONS["2.5"] = "2.5.23"