+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+- Ruby 2.6.6 is now available on heroku-22 stack via heroku-20 fallback
+
 
 ## [v313] - 2025-07-15